Added user page and more utils
